Move windows specified function to windows go files
