+++ plugins/inputs/procstat/win_service_notwindows.go
@@ -1,0 +1,11 @@
+// +build !windows
+
+package procstat
+
+import (
+	"fmt"
+)
+
+func queryPidWithWinServiceName(winServiceName string) (uint32, error) {
+	return 0, fmt.Errorf("os not support win_service option")
+}
